chatbot-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
twilio==8.11.0
google-generativeai==0.3.2
websockets==12.0
httpx==0.26.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
pydantic==2.5.3
